WIP: Utility to onboard a component
I'm really just uploading this branch as proof of work. I kinda doubt
I'll be able to get this thing in any kind of shape to be reusable.

Regarding the codecov postsubmit job:

I _think_ there should ultimately be a way to make this happen from a
config and `make jobs`. But for now...